refactor(threads): useThreadMessaging 主干外移 shared V2/V1 发送段（SharedSendContext 参数对象化，3344→2986）
